src/isf_shader_renderer/renderer.py
"""ISF shader rendering functionality."""

from pathlib import Path
from typing import Optional

from PIL import Image

from .config import Config, ShaderConfig


class ShaderRenderer:
    """Main renderer class for ISF shaders."""
    
    def __init__(self, config: Config):
        """Initialize the renderer with configuration."""
        self.config = config
    
    def render_frame(
        self,
        shader_content: str,
        time_code: float,
        output_path: Path,
        shader_config: Optional[ShaderConfig] = None,
    ) -> None:
        """
        Render a single frame of an ISF shader.
        
        Args:
            shader_content: The ISF shader source code
            time_code: The time code for this frame
            output_path: Path where the output image should be saved
            shader_config: Optional shader-specific configuration
        """
        # TODO: Implement actual ISF shader rendering
        # For now, create a placeholder image
        
        # Determine dimensions
        if shader_config:
            width = shader_config.get_width(self.config.defaults)
            height = shader_config.get_height(self.config.defaults)
            quality = shader_config.get_quality(self.config.defaults)
        else:
            width = self.config.defaults.width
            height = self.config.defaults.height
            quality = self.config.defaults.quality
        
        # Create a placeholder image
        # This will be replaced with actual shader rendering
        image = self._create_placeholder_image(shader_content, time_code, width, height)
        
        # Save the image
        output_path.parent.mkdir(parents=True, exist_ok=True)
        image.save(output_path, "PNG", optimize=True, quality=quality)
    
    def _create_placeholder_image(
        self, shader_content: str, time_code: float, width: int, height: int
    ) -> Image.Image:
        """
        Create a placeholder image for development/testing.
        
        This method will be replaced with actual ISF shader rendering.
        """
        # Create a simple gradient based on time code
        import numpy as np
        
        # Create a gradient that changes over time
        x = np.linspace(0, 1, width)
        y = np.linspace(0, 1, height)
        X, Y = np.meshgrid(x, y)
        
        # Create a time-based animation
        r = np.sin(X * 10 + time_code * 2) * 0.5 + 0.5
        g = np.cos(Y * 8 + time_code * 1.5) * 0.5 + 0.5
        b = np.sin((X + Y) * 5 + time_code * 3) * 0.5 + 0.5
        
        # Combine channels
        rgb = np.stack([r, g, b], axis=2)
        rgb = (rgb * 255).astype(np.uint8)
        
        # Convert to PIL Image
        image = Image.fromarray(rgb)
        
        return image
    
    def validate_shader(self, shader_content: str) -> bool:
        """
        Validate ISF shader content.
        
        Args:
            shader_content: The ISF shader source code
            
        Returns:
            True if the shader is valid, False otherwise
        """
        # TODO: Implement ISF shader validation
        # For now, just check if it's not empty
        return bool(shader_content.strip())
    
    def get_shader_info(self, shader_content: str) -> dict:
        """
        Extract information from ISF shader.
        
        Args:
            shader_content: The ISF shader source code
            
        Returns:
            Dictionary containing shader information
        """
        # TODO: Implement ISF shader parsing
        # For now, return basic info
        return {
            "type": "ISF",
            "size": len(shader_content),
            "lines": len(shader_content.splitlines()),
            "has_time_uniform": "TIME" in shader_content.upper(),
            "has_resolution_uniform": "RENDERSIZE" in shader_content.upper(),
        } 